coding the register
