+++ package.json
@@ -9,7 +9,6 @@
     "@testing-library/user-event": "^7.1.2",
     "antd": "^4.1.0",
     "axios": "^0.19.2",
-    "http-proxy-middleware": "^1.0.3",
     "react": "^16.13.1",
     "react-cookies": "^0.1.1",
     "react-dom": "^16.13.1",
