+++ src/index.css
@@ -1,1 +1,24 @@
-@import '../node_modules/antd/dist/antd.css';
+/* @import '../node_modules/antd/dist/antd.css';
+.site-layout-content {
+    background: #fff;
+    padding: 24px;
+    min-height: 280px;
+  }
+  .logo {
+    width: 10%;
+    height: 31px;
+    margin: 0 0 16px 0;
+    float: left;
+  }
+  .search{
+    width: 30%;
+    height: 30px;
+    margin: 16px 20px 10px 2px;
+    float: left;
+  }
+  .menuItemStyle{
+    width: 10%;
+    text-align: center;
+    margin-left: 20px;
+    margin-right: 20px;
+  } */
